requirements.txt
transformers
datasets
sentence-transformers
scikit-learn
python-dotenv
pandas
torch
